fix links on mobile
redirect mobile to the view doc mode
open header links externally:
that will open in the GDocs mobile app

closes #136
